requirements.txt
DiscordUtils
aiogoogletrans
alexflipnote.py
async_cleverbot
async_cse
asyncpg
bs4
dblpy
git+git://github.com/Rapptz/discord.py.git
git+git://github.com/ShineyDev/prettify.py.git
gTTS
git+git://github.com/jmoiron/humanize.git
jishaku
lxml
numexpr
psutil
pyparsing
randomcolor
requests
tabulate
uwuify
wikipedia
youtube_dl